==> 3_CalculateCovariates.Rmd <==
---
title: "CalculateCovariates"
author: "Amy Van Scoyoc"
date: "9/1/2022"
output: github_document
---

# Covariate Calculations

```{r}
library(tidyverse)
library(raster)
library(sf)
#install.packages("exactextractr")
library(exactextractr)
```

```{r}
cameras_shp <- st_read("data/camera_locations.shp") #import camera site locations
```


```{r}
setwd("data/covariates/")

elevation <- raster("SL_elevation.tif")
slope <- raster("SL_slope.tif")
# water <- raster("SL_water.tif")
# hf <- raster("SL_humanfootprint.tif")
d_corral <- raster("d_corral.tif")
ndvi <- raster("SL_NDVI.tiff")
d_setmt <- raster("d_settlement.tif")
prey <- read_csv("prey_rai.csv")
```
## For distance to corrals

Custom function for calculating near distances
```{r}
# Calculate_Near <- function(input_feature, near_feature) {
#   Near_index <- st_nearest_feature(input_feature, near_feature)
#   sliced <- near_feature %>% slice(Near_index)
#   near_distance <- st_distance(input_feature, sliced, by_element = TRUE)
#   return(near_distance)
# }
```

Calculations
```{r}
# covariates <- cameras_shp %>%
#   mutate(dist_corrals = Calculate_Near(cameras_shp, corrals),
#          elev = raster::extract(elevation, .),
#          slope = raster::extract(slope, .))
```

Livestock Density
```{r}
record_table <- read.csv("data/Data-21-22_Record_Table.csv") 

livestock_rt <- record_table %>% filter(Species == "Livestock") 
```


```{r}
buffers100 <- cameras_shp %>% st_buffer(dist = 100) #create buffers around cameras

buffers100_extract <- buffers100 %>% #extract avg values from buffered area
                mutate(d_corral = exactextractr::exact_extract(d_corral, buffers100, 'median'),
                       d_setmt = exactextractr::exact_extract(d_setmt, buffers100, 'median'), 
                       ndvi = exactextractr::exact_extract(ndvi, buffers100, 'median'),
                       elev = raster::extract(elevation, cameras_shp),
                       slope = raster::extract(slope, cameras_shp)) %>% 
                st_drop_geometry() 

covariates <- buffers100_extract %>% 
              left_join(prey) %>% 
              mutate(lvstk_occ = c(0,0,0,0,0,0,0,0,1,1,0,0,0,0,1,0,1,0,0,0,0,0,0,0,0,1,0,0,0,0))
covariates
```


### Write to covariates to camera location file
```{r}
write.csv(covariates, "data/covariates.csv")
```
