==> 4_Occupancy_SSOM.Rmd <==
---
title: "Occupancy_SSOM"
author: "Amy Van Scoyoc"
date: "9/1/2022"
output: github_document
---

# Libraries

```{r}
library(tidyverse) # for basic data wrangling
library(camtrapR) # for creating species record tables
```

```{r}
record_table <- read.csv("data/Data-21-22_Record_Table.csv") #import record table
camtraps <- read.csv("data/cameratraps.csv") #import camera site locations
```

```{r}
#prey
record_table[record_table$Species == "Ibex" | record_table$Species == "Markhor",] %>% write.csv("data/record_table_prey.csv")
```


```{r}
# spatial location of detections

#subset record table
record_table_SL <- record_table[record_table$Species == "Snow Leopard",]
 
SL_detections <- detectionMaps(CTtable   = camtraps,
                           recordTable   = record_table_SL,
                           Xcol          = "longitude",
                           Ycol          = "latitude",
                           stationCol    = "Station",
                           speciesCol    = "Species",
                           speciesToShow = "Snow Leopard", #add if number detections
                           printLabels   = FALSE,
                           richnessPlot  = FALSE, #T if richness for all species
                           speciesPlots  = TRUE, #T if number of detections for one species
                           addLegend     = TRUE)
```


```{r}
# create camera operation matrix
camop_no_problem <- cameraOperation(CTtable      = camtraps,
                                    stationCol   = "Station",
                                    setupCol     = "set_date",
                                    retrievalCol = "close_date",
                                    hasProblems  = FALSE,
                                    dateFormat   = "%Y-%m-%d")
```


```{r}
# create detection history (with trapping effort)
det_hist_sl <- detectionHistory(recordTable         = record_table,
                            camOp                = camop_no_problem,
                            stationCol           = "Station",
                            speciesCol           = "Species",
                            recordDateTimeCol    = "DateTimeOriginal",
                            species              = "Snow Leopard",
                            occasionLength       = 7,
                            day1                 = "station",
                            includeEffort        = T, #T if trap effort influences detection
                            scaleEffort          = TRUE) #to help with convergence of models

det_hist_sl <- as.data.frame(det_hist_sl) 
```

# Occupancy model in unmarked

```{r}
library(unmarked)
```

### Site covariates
```{r}
library(here)
covariates <- read_csv(here("data/covariates.csv"))

site <- data.frame(elev = covariates$elvtn_m,
                   slope = covariates$slope,
                   d_corral = covariates$d_corral, 
                   d_setmt = covariates$d_setmt,
                   ndvi = covariates$ndvi,
                   prey = covariates$prey_rai,
                   lvstk_occ = as.factor(covariates$lvstk_occ),
                   n_days = covariates$n_days)
```

### Detection covariates
```{r}
obs <- 
```

### Create unmarked dataframe
```{r}
sl_occ <- unmarkedFrameOccu(y = det_hist[,c(1:21)], siteCovs = site)
summary(sl_occ)
``` 

### Run occupancy model
```{r}
cor(covariates[,4:12])

model1 = occu(~n_days + elev ~d_corral + prey + lvstk_occ, data = sl_occ, se = T)

coef(model1, type = "state")
summary(model1)
```

### Backtransform
```{r}
backTransform(linearComb(model1, coefficients = c(1,1),"state")) #update coeff & "state/det"
```

### Plot
```{r}
# Plotting habitat occupancy probability
newdata1 <- data.frame(siteCovs=c("Edge", "Interior"), date = 0,dates.2 = 0,mins = 0,mins.2 = 0)

#Make simulated data for the predict function 
simdat <- predict(model1, type="state", newdata=newdata1)

with(simdat,{
  plot(1:2, Predicted, xaxt="n", xlim=c(0.5, 2.5),ylim = c(0,1),
       xlab="Habitat",
       ylab=expression(paste("Probability of occurrence (", psi, ")")),
       cex.lab=1.2,
       pch=16, cex=1.5, main = "Probability of Occupancy for Each Habitat")
  axis(1, 1:2, c('Interior', 'Edge'))
  arrows(1:2, lower, 1:2, upper, angle=90, code=3, length=0.05)})
```
